% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{A}
\alias{A}
\title{Amino Acid Analysis Function}
\usage{
A(Locus, loci.ColNames, genos, grp, Strict.Bin, ExonAlign, Cores)
}
\arguments{
\item{Locus}{Locus being analyzed.}

\item{loci.ColNames}{The column names of the loci being analyzed.}

\item{genos}{Genotype table.}

\item{grp}{Case/Control or Phenotype groupings.}

\item{Strict.Bin}{Logical specify if strict rare cell binning should be used in ChiSq test.}

\item{ExonAlign}{Exon protein alignment filtered for locus.}

\item{Cores}{Number of cores to use for analysis.}
}
\description{
This is the workhorse function for the amino acid analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
