% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_wrapper.R
\name{A.wrapper}
\alias{A.wrapper}
\title{Amino Acid Wrapper}
\usage{
A.wrapper(
  loci,
  loci.ColNames,
  genos,
  grp,
  nGrp0,
  nGrp1,
  EPL,
  Cores,
  Output,
  Verbose
)
}
\arguments{
\item{loci}{Loci being analyzed.}

\item{loci.ColNames}{The column names of the loci being analyzed.}

\item{genos}{Genotype table.}

\item{grp}{Case/Control or Phenotype groupings.}

\item{nGrp0}{Number of controls.}

\item{nGrp1}{Number of cases.}

\item{EPL}{Exon protein alignment.}

\item{Cores}{Number of cores to use for analysis.}

\item{Output}{Data return carryover from main BIGDAWG function}

\item{Verbose}{Summary display carryover from main BIGDAWG function}
}
\description{
Wrapper function for amino acid analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
