% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_support_functions.R
\name{AA.df.check}
\alias{AA.df.check}
\title{Contingency Table Check}
\usage{
AA.df.check(x, Strict.Bin)
}
\arguments{
\item{x}{contingency table.}

\item{Strict.Bin}{Logical specify if strict rare cell binning should be used in ChiSq test.}
}
\description{
Checks amino acid contingency table data frame to ensure required variation exists.
}
\note{
This function is for internal BIGDAWG use only.
}
