% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_support_functions.R
\name{AA.df.cs}
\alias{AA.df.cs}
\title{Contingency Table Amino Acid ChiSq Testing}
\usage{
AA.df.cs(x, Strict.Bin)
}
\arguments{
\item{x}{contingency table.}

\item{Strict.Bin}{Logical specify if strict rare cell binning should be used in ChiSq test.}
}
\description{
Runs ChiSq test on amino acid contingency table data frames.
}
\note{
This function is for internal BIGDAWG use only.
}
