% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_support_functions.R
\name{AAtable.builder}
\alias{AAtable.builder}
\title{Amino Acid Contingency Table Build}
\usage{
AAtable.builder(x, y)
}
\arguments{
\item{x}{Filtered alignmnet list element.}

\item{y}{Phenotype groupings.}
}
\description{
Build Contingency Tables for Amino Acid Analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
