% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_support_functions.R
\name{AlignmentFilter}
\alias{AlignmentFilter}
\title{Alignment Filter}
\usage{
AlignmentFilter(Align, Alleles, Locus)
}
\arguments{
\item{Align}{Protein Alignment Object.}

\item{Alleles}{to be pulled.}

\item{Locus}{Locus to be filtered against.}
}
\description{
Filter Protein Exon Alignment File for Specific Alleles.
}
\note{
This function is for internal BIGDAWG use only.
}
