% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_ExonPtnAlign_functions.R
\name{AlignObj.Update}
\alias{AlignObj.Update}
\title{Updated Alignment Object Creator}
\usage{
AlignObj.Update(Loci, Release, RefTab)
}
\arguments{
\item{Loci}{Loci to be bundled.}

\item{Release}{IMGT/HLA database release version.}

\item{RefTab}{Data of reference exons used for protein alignment creation.}
}
\description{
Synthesize Object for Exon Protein Alignments.
}
\note{
This function is for internal BIGDAWG use only.
}
