% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Append.System}
\alias{Append.System}
\title{Append Genetic System Locus Designation to Allele String}
\usage{
Append.System(x, df.name)
}
\arguments{
\item{x}{Vector Column genotypes to append}

\item{df.name}{String SystemLocus name for each allele.}
}
\description{
Adds genetic system (HLA/KIR) to each allele name
}
\note{
This function is for internal use only.
}
