% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Build.Matrix}
\alias{Build.Matrix}
\title{Build Output Matrix for GL2Tab Conversion}
\usage{
Build.Matrix(System, Loci)
}
\arguments{
\item{System}{Character Genetic system HLA- or KIR}

\item{Loci}{The loci for header names}
}
\description{
Initializes output matrix format for GL2Tab conversion
}
\note{
This function is for internal use only.
}
