% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_support_functions.R
\name{buildHAPnames}
\alias{buildHAPnames}
\title{Haplotype Name Builder}
\usage{
buildHAPnames(Combn, loci)
}
\arguments{
\item{Combn}{Combination of loci to extraction from genos}

\item{loci}{Character vector of unique loci being analyzed.}
}
\description{
Builds table of names for HAPsets
}
\note{
This function is for internal BIGDAWG use only.
}
