% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{cci.pval}
\alias{cci.pval}
\title{Case Control Odds Ratio Calculation from Epicalc}
\usage{
cci.pval(x)
}
\arguments{
\item{x}{List of 2x2 matrices for calculation, output of TableMaker.}
}
\description{
Calculates odds ratio and pvalues from 2x2 table
}
\note{
This function is for internal BIGDAWG use only.
}
