% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{Check.Cores}
\alias{Check.Cores}
\title{Check Cores Parameters}
\usage{
Check.Cores(Cores.Lim, Output)
}
\arguments{
\item{Cores.Lim}{Integer How many cores can be used.}

\item{Output}{Logical Should analysis results be written to output directory.}
}
\description{
Check cores limitation for OS compatibility
}
