% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{Check.Data}
\alias{Check.Data}
\title{Check Data Structure}
\usage{
Check.Data(Data, Convert)
}
\arguments{
\item{Data}{String Type of output.}

\item{Convert}{String Direction for conversion.}
}
\description{
Check data structure for successfuly conversion.
}
\note{
This function is for internal use only.
}
