% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{CheckAlleles}
\alias{CheckAlleles}
\title{Allele Legitimacy Check}
\usage{
CheckAlleles(x, y, z1, z2)
}
\arguments{
\item{x}{Exon protein list alignment object.}

\item{y}{Genotypes from data file}

\item{z1}{loci in data file}

\item{z2}{Genotype column names}
}
\description{
Checks available alleles against data to ensure complete overlap.
}
\note{
This function is for internal BIGDAWG use only.
}
