% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{CheckHLA}
\alias{CheckHLA}
\title{HLA Formatting Check for Amino Acid Analysis}
\usage{
CheckHLA(x)
}
\arguments{
\item{x}{All columns of HLA genotyping data.}
}
\description{
Checks data to see if HLA data is properly formatted .
}
\note{
This function is for internal BIGDAWG use only.
}
