% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{CheckLoci}
\alias{CheckLoci}
\title{HLA Loci Legitimacy Check for Amino Acid Analysis}
\usage{
CheckLoci(x, y)
}
\arguments{
\item{x}{Loci available in exon protein list alignment object.}

\item{y}{Unique column names}
}
\description{
Checks available loci against data to ensure complete overlap.
}
\note{
This function is for internal BIGDAWG use only.
}
