% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{CheckRelease}
\alias{CheckRelease}
\title{Function to Check Release Versions}
\usage{
CheckRelease(Package = T, Alignment = T, Output = F)
}
\arguments{
\item{Package}{Logical to check for BIGDAWG package versions}

\item{Alignment}{Logical to check the IMGT/HLA database version for the alignment bundled with BIGDAWG.}

\item{Output}{Should any error be written to a file}
}
\description{
This updates the protein aligment used in checking HLA loci and alleles as well as in the amino acid analysis.
}
\note{
Requires active internet connection.
}
