% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{CheckString.Allele}
\alias{CheckString.Allele}
\title{GL String Allele Check}
\usage{
CheckString.Allele(x)
}
\arguments{
\item{x}{GL String to check against}
}
\description{
GL String check for allele ambiguity formatting
}
\note{
This function is for internal use only.
}
