% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{CheckString.Locus}
\alias{CheckString.Locus}
\title{GL String Locus Check}
\usage{
CheckString.Locus(x, Loci)
}
\arguments{
\item{x}{GL String to check against}

\item{Loci}{Loci to check}
}
\description{
Check GL string for loci appearing in multiple gene fields.
}
\note{
This function is for internal use only.
}
