% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_support_functions.R
\name{Condense.EPL}
\alias{Condense.EPL}
\title{Condensing Exon Specific Alignments to Single Dataframe}
\usage{
Condense.EPL(EPL.Exon)
}
\arguments{
\item{EPL.Exon}{Exon-Locus Specific Amino Acid Alignment.}
}
\description{
Combines multiple Exon Specific Alignments into a single Alignment object
}
\note{
This function is for internal BIGDAWG use only.
}
