% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_support_functions.R
\name{Create.Null.Table}
\alias{Create.Null.Table}
\title{Create Empty Table}
\usage{
Create.Null.Table(Locus, Names, nr)
}
\arguments{
\item{Locus}{Locus being analyzed.}

\item{Names}{Column names for final matrix.}

\item{nr}{Number of rows.}
}
\description{
Creates matrix of NA for no result tables.
}
\note{
This function is for internal BIGDAWG use only.
}
