% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRB_Parser.R
\name{DRB345.Check.Wrapper}
\alias{DRB345.Check.Wrapper}
\title{DRB345 haplotype zygosity wrapper}
\usage{
DRB345.Check.Wrapper(Genotype, Loci.DR)
}
\arguments{
\item{Genotype}{Row of data set data frame following DRB345 parsing}

\item{Loci.DR}{DRBx Loci of interest to test for consistency}
}
\description{
Checks DR haplotypes for correct zygosity and flags unanticipated haplotypes
}
\note{
This function is for internal use only.
}
