% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRB_Parser.R
\name{DRB345.Check.Zygosity}
\alias{DRB345.Check.Zygosity}
\title{DRB345 haplotype zygosity checker single locus}
\usage{
DRB345.Check.Zygosity(Locus, Genotype)
}
\arguments{
\item{Locus}{Locus of interest to test for consistency}

\item{Genotype}{Row of data set data frame following DRB345 parsing}
}
\description{
Checks DR haplotypes for correct zygosity and flags unanticipated haplotypes for a single DRBx
}
\note{
This function is for internal use only.
}
