% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRB_Parser.R
\name{DRB345.Exp}
\alias{DRB345.Exp}
\title{DRB345 Expected}
\usage{
DRB345.Exp(DRB1.Genotype)
}
\arguments{
\item{DRB1.Genotype}{DRB1 Subject Genotypes}
}
\description{
Checks DRB1 Genotype and Returns Expected DR345 Loci
}
\note{
This function is for internal use only.
}
