% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRB_Parser.R
\name{DRB345.flag}
\alias{DRB345.flag}
\title{DRB345 haplotype zygosity flag check}
\usage{
DRB345.flag(tmp, Tab)
}
\arguments{
\item{tmp}{Output of DRB345.zygosity}

\item{Tab}{Data frame of sampleIDs, phenotypes, and genotypes}
}
\description{
Identify DR345 flagged haplotypes
}
\note{
This function is for internal BIGDAWG use only.
}
