% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRB_Parser.R
\name{DRB345.parser}
\alias{DRB345.parser}
\title{DRB345 Column Processing}
\usage{
DRB345.parser(Tab)
}
\arguments{
\item{Tab}{Data frame of sampleIDs, phenotypes, and genotypes}
}
\description{
Separates DRB345 column pair into separate columns for each locus
}
\note{
This function is for internal BIGDAWG use only.
}
