% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRB_Parser.R
\name{DRB345.zygosity}
\alias{DRB345.zygosity}
\title{DRB345 haplotype zygosity checker}
\usage{
DRB345.zygosity(x)
}
\arguments{
\item{x}{Row of data set data frame following DRB345 parsing}
}
\description{
Checks DR haplotypes for correct zygosity and flags unanticipated haplotypes
}
\note{
This function is for internal BIGDAWG use only.
}
