% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrLog.R
\name{Err.Log}
\alias{Err.Log}
\title{Error Code Display and Logging}
\usage{
Err.Log(Output, x, y = NULL, z = NULL)
}
\arguments{
\item{Output}{Logical indicating if Error logging should be written to a file.}

\item{x}{Log Code.}

\item{y}{Misc information relevant to error.}

\item{z}{Misc information relevant to error.}
}
\description{
Displays error codes attributable to data formatting and Locus/Allele naming. Writes to log file.
}
\note{
This function is for internal BIGDAWG use only.
}
