% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{EVSremoval}
\alias{EVSremoval}
\title{Expression Variant Suffix Removal}
\usage{
EVSremoval(Locus, EPList)
}
\arguments{
\item{Locus}{Locus to be filtered against.}

\item{EPList}{Exon Protein Alignment Object}
}
\description{
Removes expression variant suffixes from HLA alleles in the exon protein alignment object.
}
\note{
This function is for internal BIGDAWG use only.
}
