% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_support_functions.R
\name{Exon.Filter}
\alias{Exon.Filter}
\title{Filter Exon Specific Alignment Sections}
\usage{
Exon.Filter(Locus, Exon, EPL.Locus, RefExons, E.Ptn.Starts)
}
\arguments{
\item{Locus}{Locus being analyzed.}

\item{Exon}{Exon being analyzed.}

\item{EPL.Locus}{ExonPtnAlign object filtered by Locus}

\item{RefExons}{Reference Exon Table}

\item{E.Ptn.Starts}{Exon Protein Overlay Map}
}
\description{
Filters the ExonPtnAlign object by locus and exon.
}
\note{
This function is for internal BIGDAWG use only.
}
