% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_ExonPtnAlign_functions.R
\name{ExonPtnAlign.Create}
\alias{ExonPtnAlign.Create}
\title{Protein Exon Alignment Formatter}
\usage{
ExonPtnAlign.Create(Locus, RefTab)
}
\arguments{
\item{Locus}{Locus alignment to be formatted.}

\item{RefTab}{Reference exon protein information for alignment formatting.}
}
\description{
Dynamically creates an alignmnet of Allele exons for Analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
