% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExonPtnList}
\alias{ExonPtnList}
\title{Exon 2 (class II) or 2/3 (class I) protein alignments.}
\format{
A list wherein each element is an alignment dataframe for a single locus.
}
\usage{
ExonPtnList
}
\description{
Alignment object for use in the amino acid analysis.
}
\keyword{datasets}
