% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Filler}
\alias{Filler}
\title{Replace or Fill 00:00 allele strings}
\usage{
Filler(x, Locus = NULL, Type)
}
\arguments{
\item{x}{Genotype}

\item{Locus}{Locus column to adjust.}

\item{Type}{String specifying whether to pad ('Fill') or leave blank ('Remove') absent calls}
}
\description{
Replaces or Fills absent allele strings.
}
\note{
This function is for internal use only.
}
