% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{FixAlleleName}
\alias{FixAlleleName}
\title{Allele Name Format Fix}
\usage{
FixAlleleName(Output, Tab)
}
\arguments{
\item{Output}{Logical indicating if Error logging should be written to a file.}

\item{Tab}{All columns of HLA genotyping data.}
}
\description{
Separate locus and allele names if data is formatted as Loci*Allele
}
\note{
This function is for internal BIGDAWG use only.
}
