% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Format.Allele}
\alias{Format.Allele}
\title{Ambiguous Alleles Locus Name Formatting}
\usage{
Format.Allele(x, Type)
}
\arguments{
\item{x}{Allele String}

\item{Type}{String specifying whether to strip ('off') or append ('on') locus prefix}
}
\description{
Remove or Append Locus name from/to allele in an ambiguous allele string
}
\note{
This function is for internal use only.
}
