% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Format.Tab}
\alias{Format.Tab}
\title{Tabular Data Locus Format Tool}
\usage{
Format.Tab(x, Order)
}
\arguments{
\item{x}{Single row of converted GL string}

\item{Order}{Single row data frame for mapping converted GL strings}
}
\description{
Correctly orders the expanded GL string
}
\note{
This function is for internal use only.
}
