% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE_support_functions.R
\name{getAllele.Count}
\alias{getAllele.Count}
\title{Recompute number of alleles}
\usage{
getAllele.Count(x)
}
\arguments{
\item{x}{Locus specific contingency matrix getCS.Mat output.}
}
\description{
Using Freq.Final, recompute number of alleles
}
\note{
This function is for internal BIGDAWG use only.
}
