% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE_support_functions.R
\name{getCS.Mat}
\alias{getCS.Mat}
\title{Chi square matrices}
\usage{
getCS.Mat(Locus, genos.sub, Allele.Freq, Allele.Combn)
}
\arguments{
\item{Locus}{Locus of interest.}

\item{genos.sub}{Genotypes for locus of interest.}

\item{Allele.Freq}{Allele frequencies.}

\item{Allele.Combn}{Allele combinations.}
}
\description{
Chi Square contingency matrix builder with rare cell binning
}
\note{
This function is for internal BIGDAWG use only.
}
