% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE_support_functions.R
\name{getCS.stat}
\alias{getCS.stat}
\title{Chi square test statistic}
\usage{
getCS.stat(Locus, Freq.Final)
}
\arguments{
\item{Locus}{Locus of interest.}

\item{Freq.Final}{Contingency Matrix getCS.Mat output.}
}
\description{
Calculate chi square test statistic
}
\note{
This function is for internal BIGDAWG use only.
}
