% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{GetField}
\alias{GetField}
\title{HLA trimming function}
\usage{
GetField(x, Res)
}
\arguments{
\item{x}{HLA allele.}

\item{Res}{Resolution desired.}
}
\description{
Trim a properly formatted HLA allele to desired number of fields.
}
\note{
This function is for internal BIGDAWG use only.
}
