% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{getFileName}
\alias{getFileName}
\title{File Name Extraction}
\usage{
getFileName(x)
}
\arguments{
\item{x}{File name.}
}
\description{
Function to extract file path.
}
\note{
This function is for internal use only.
}
