% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_ExonPtnAlign_functions.R
\name{GetFiles}
\alias{GetFiles}
\title{File Fetcher}
\usage{
GetFiles(Loci)
}
\arguments{
\item{Loci}{HLA Loci to be fetched. Limited Loci available.}
}
\description{
Download Protein Alignment and Accessory Files
}
\note{
This function is for internal BIGDAWG use only.
}
