% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_support_functions.R
\name{getHap}
\alias{getHap}
\title{Haplotype Table Maker}
\usage{
getHap(SID, HaploEM)
}
\arguments{
\item{SID}{Index number (i.e., row number) of sample ID from genotype matrix.}

\item{HaploEM}{Haplotype output object from haplo.stat::haplo.em function.}
}
\description{
Builds table of haplotypes
}
\note{
This function is for internal BIGDAWG use only.
}
