% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE_support_functions.R
\name{getObsFreq}
\alias{getObsFreq}
\title{Observed Frequency}
\usage{
getObsFreq(x, genos.locus)
}
\arguments{
\item{x}{Single genotype.}

\item{genos.locus}{Locus genotypes.}
}
\description{
Get observed frequency of genotypes
}
\note{
This function is for internal BIGDAWG use only.
}
