% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL2TAB.R
\name{GL2Tab.Loci}
\alias{GL2Tab.Loci}
\title{Locus Ordering for GL2Tab}
\usage{
GL2Tab.Loci(Locus, Genotype, System)
}
\arguments{
\item{Locus}{Locus to condense}

\item{Genotype}{Row of loci to condense}

\item{System}{Character Genetic system HLA or KIR}
}
\description{
Orders Locus Calls
}
\note{
This function is for internal use only.
}
