% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL2TAB.R
\name{GL2Tab.Sub}
\alias{GL2Tab.Sub}
\title{Genotype List String Expander}
\usage{
GL2Tab.Sub(x, System)
}
\arguments{
\item{x}{Character GL string to expand}

\item{System}{Character Genetic system HLA or KIR}
}
\description{
Expands GL string into a table of adjacent loci
}
\note{
This function is for internal use only.
}
