% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL2TAB.R
\name{GL2Tab.wrapper}
\alias{GL2Tab.wrapper}
\title{Genotype List String to Tabular Data Conversion}
\usage{
GL2Tab.wrapper(df, System, Strip.Prefix, Abs.Fill, Cores)
}
\arguments{
\item{df}{Data frame containing GL strings}

\item{System}{Character Genetic system HLA or KIR}

\item{Strip.Prefix}{Logical Should System/Locus prefixes be stripped from table data.}

\item{Abs.Fill}{Logical Should absent loci special designations be used.}

\item{Cores}{Integer How many cores can be used}
}
\description{
Expands GL strings to columns of adjacent locus pairs.
}
\note{
This function is for internal use only
}
