% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLSconvert.R
\name{GLSconvert}
\alias{GLSconvert}
\title{Genotype List String Conversion}
\usage{
GLSconvert(
  Data,
  Convert,
  File.Output = "txt",
  System = "HLA",
  HZY.Red = FALSE,
  DRB345.Check = FALSE,
  Strip.Prefix = TRUE,
  Abs.Fill = FALSE,
  Cores.Lim = 1L
)
}
\arguments{
\item{Data}{String File name or R Data Frame.}

\item{Convert}{String Direction for conversion.}

\item{File.Output}{String Type of File.Output.}

\item{System}{String Genetic system (HLA or KIR) of the data being converted}

\item{HZY.Red}{Logical Reduction of homozygote genotypes to single allele.}

\item{DRB345.Check}{Logical Check DR haplotypes for consistency and flag unusual haplotypes.}

\item{Strip.Prefix}{Logical Should System/Locus prefixes be stripped from table data.}

\item{Abs.Fill}{Logical Should absent loci special designations be used.}

\item{Cores.Lim}{Integer How many cores can be used.}
}
\description{
Main Workhorse wrapper for cross converting columnar table to GL string representaion.
}
