% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_wrapper_MC.R
\name{H.MC.wrapper}
\alias{H.MC.wrapper}
\title{Haplotype Wrapper for Multicore}
\usage{
H.MC.wrapper(
  SID,
  Tabsub,
  loci,
  loci.ColNames,
  genos,
  grp,
  All.Pairwise,
  Strict.Bin,
  Output,
  Verbose,
  Cores
)
}
\arguments{
\item{SID}{Character vector of subject IDs.}

\item{Tabsub}{Data frame of genotype calls for set being analyzed.}

\item{loci}{Character vector of unique loci being analyzed.}

\item{loci.ColNames}{Character vector of genos column names.}

\item{genos}{The genotype columns of the loci set being analyzed.}

\item{grp}{Case/Control or Phenotype groupings.}

\item{All.Pairwise}{Haplotype argument carryover from main BIGDAWG function}

\item{Strict.Bin}{Logical specify if strict rare cell binning should be used in ChiSq test}

\item{Output}{Data return carryover from main BIGDAWG function}

\item{Verbose}{Summary display carryover from main BIGDAWG function}

\item{Cores}{Cores carryover from main BIGDAWG function}
}
\description{
Wrapper for main H function
}
\note{
This function is for internal BIGDAWG use only.
}
