% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.R
\name{H}
\alias{H}
\title{Haplotype Analysis Function}
\usage{
H(genos.sub, grp)
}
\arguments{
\item{genos.sub}{The genotype columns of the loci(locus) set being analyzed.}

\item{grp}{Case/Control or Phenotype groupings.}
}
\description{
This is the workhorse function for the haplotype analysis.
}
\note{
This function is for internal BIGDAWG use only.
}

