% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLA_data}
\alias{HLA_data}
\title{Example HLA Dataset}
\format{
A data frame with 2000 rows and 14 variables
}
\usage{
HLA_data
}
\description{
A synthetic dataset of HLA genotypes for using bigdawg.
}
\keyword{datasets}
