% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE_support_functions.R
\name{HWE.ChiSq}
\alias{HWE.ChiSq}
\title{Hardy Weinbergy Equilibrium Function}
\usage{
HWE.ChiSq(genos.sub, loci, nloci)
}
\arguments{
\item{genos.sub}{data frame of genotype files post processing.}

\item{loci}{list of loci.}

\item{nloci}{number of loci in list}
}
\description{
This is the workhorse function for each group analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
