% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE_wrapper.R
\name{HWE.wrapper}
\alias{HWE.wrapper}
\title{Hardy-Weinbery Wrapper}
\usage{
HWE.wrapper(Tab, Output, Verbose)
}
\arguments{
\item{Tab}{Data frame of genotype files post processing.}

\item{Output}{Data return carryover from main BIGDAWG function}

\item{Verbose}{Summary display carryover from main BIGDAWG function}
}
\description{
Wrapper for main HWE function
}
\note{
This function is for internal BIGDAWG use only.
}
