% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HWE.R
\name{HWEChiSq}
\alias{HWEChiSq}
\title{Hardy Weinbergy Equilibrium Function}
\usage{
HWEChiSq(Tab, All.ColNames)
}
\arguments{
\item{Tab}{data frame of genotype files post processing.}

\item{All.ColNames}{character vector of Tab object column names.}
}
\description{
This is the main wrapper function for each analysis.
}
\note{
This function is for internal BIGDAWG use only.
}

