% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L.R
\name{L}
\alias{L}
\title{Locus Analysis Function 25Jun2025}
\usage{
L(loci.ColNames, Locus, genos, grp, Strict.Bin)
}
\arguments{
\item{loci.ColNames}{The column names of the loci being analyzed.}

\item{Locus}{Locus being analyzed.}

\item{genos}{Genotype table}

\item{grp}{Case/Control or Phenotype groupings.}

\item{Strict.Bin}{Logical specify if strict rare cell binning should be used in ChiSq test}
}
\description{
This is the workhorse function for the locus level analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
