% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_wrapper.R
\name{L.wrapper}
\alias{L.wrapper}
\title{Locus Wrapper}
\usage{
L.wrapper(nloci, loci, loci.ColNames, genos, grp, Strict.Bin, Output, Verbose)
}
\arguments{
\item{nloci}{Number of loci being analyzed.}

\item{loci}{Loci being analyzed.}

\item{loci.ColNames}{The column names of the loci being analyzed.}

\item{genos}{Genotype table}

\item{grp}{Case/Control or Phenotype groupings.}

\item{Strict.Bin}{Logical specify if strict rare cell binning should be used in ChiSq test}

\item{Output}{Data return carryover from main BIGDAWG function}

\item{Verbose}{Summary display carryover from main BIGDAWG function}
}
\description{
Wrapper for main L function
}
\note{
This function is for internal BIGDAWG use only.
}
