% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{make2x2}
\alias{make2x2}
\title{Creation of a 2x2 table using the indicated orientation.}
\usage{
make2x2(caseexp, controlex, casenonex, controlnonex)
}
\arguments{
\item{caseexp}{Number of cases exposed}

\item{controlex}{Number of controls exposed}

\item{casenonex}{Number of cases not exosed}

\item{controlnonex}{Number of controls not exposed}
}
\description{
make2x2 function port epicalc version 2.15.1.0 (Virasakdi Chongsuvivatwong, 2012)
}
\note{
This function is for internal BIGDAWG use only.
}
