% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE_support_functions.R
\name{makeComb}
\alias{makeComb}
\title{Genotype Combination Maker}
\usage{
makeComb(x)
}
\arguments{
\item{x}{Number of alleles.}
}
\description{
Make data frame of possible genotype combinations
}
\note{
This function is for internal BIGDAWG use only.
}
