% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{MergeData_Output}
\alias{MergeData_Output}
\title{Data Object Merge and Output}
\usage{
MergeData_Output(BD.out, Run, OutDir)
}
\arguments{
\item{BD.out}{Output of analysis as list.}

\item{Run}{Tests that are to be run as defined by Run.Tests.}

\item{OutDir}{Output directory defined by Results.Dir or default.}
}
\description{
Whole data set table construction of per haplotype for odds ratio, confidence intervals, and pvalues
}
\note{
This function is for internal BIGDAWG use only.
}
