% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_ExonPtnAlign_functions.R
\name{PgrpExtract}
\alias{PgrpExtract}
\title{HLA P group Finder}
\usage{
PgrpExtract(x, y)
}
\arguments{
\item{x}{Allele of interest.}

\item{y}{Formatted P groups.}
}
\description{
Identify P group for a given allele if exists.
}
\note{
This function is for internal BIGDAWG use only.
}
