% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_ExonPtnAlign_functions.R
\name{PgrpFormat}
\alias{PgrpFormat}
\title{HLA P group File Formatter}
\usage{
PgrpFormat(x, Locus)
}
\arguments{
\item{x}{P group object from read.table command.}

\item{Locus}{Locus to be filtered on.}
}
\description{
Format the hla_nom_p.txt read table object for a specific locus.
}
\note{
This function is for internal BIGDAWG use only.
}
