% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_functions.R
\name{PreCheck}
\alias{PreCheck}
\title{Data summary function}
\usage{
PreCheck(Tab, All.ColNames, rescall, HLA)
}
\arguments{
\item{Tab}{Loci available in exon protein list alignment object.}

\item{All.ColNames}{Column names from genotype data.}

\item{rescall}{HLA resolution set for analysis.}

\item{HLA}{HLA bigdawg argument passed to function}
}
\description{
Summary function for sample population within data file.
}
\note{
This function is for internal BIGDAWG use only.
}

