% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{prepData}
\alias{prepData}
\title{Prepare imported data v 1.1 10 December 2025}
\usage{
prepData(Tab, data_type)
}
\arguments{
\item{Tab}{Genotypes dataframe.}

\item{data_type}{The type of data being processed. Options are 'Locus' or 'Motif'.'}
}
\description{
Prepare imported data for processing, checks, and analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
