% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{rmABstrings}
\alias{rmABstrings}
\title{Replace absent allele strings}
\usage{
rmABstrings(df)
}
\arguments{
\item{df}{Genotypes dataframe.}
}
\description{
Replaces allowable absent allele strings with ^ symbol.
}
\note{
This function is for internal BIGDAWG use only.
}
