% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{RunChiSq_c}
\alias{RunChiSq_c}
\title{Contextual Binning Chi-squared Contingency Table Test}
\usage{
RunChiSq_c(x)
}
\arguments{
\item{x}{Contingency table.}
}
\description{
Calculates chi-squared contingency table tests and bins rare cells at 20% capture rate.
}
\note{
This function is for internal BIGDAWG use only.
}
