% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Stripper}
\alias{Stripper}
\title{Removes System and Locus from Alleles}
\usage{
Stripper(x)
}
\arguments{
\item{x}{Allele}
}
\description{
Removes the System and Locus designations for alleles calls in GL2Tab
}
\note{
This function is for internal use only.
}
