% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{summaryGeno.2}
\alias{summaryGeno.2}
\title{Haplotype missing Allele summary function}
\usage{
summaryGeno.2(geno, miss.val = 0)
}
\arguments{
\item{geno}{Matrix of genotypes.}

\item{miss.val}{Vector of codes for allele missing values.}
}
\description{
Summary function for identifying missing alleles in a matrix of genotypes.
}
\note{
This function is for internal BIGDAWG use only and is ported from haplo.stats.
}
