% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAB2GL.R
\name{Tab2GL.Loci}
\alias{Tab2GL.Loci}
\title{Locus Condenser for Tab2GL}
\usage{
Tab2GL.Loci(Locus, Genotype, System, HZY.Red)
}
\arguments{
\item{Locus}{Locus to condense}

\item{Genotype}{Row of loci to condense}

\item{System}{Character Genetic system HLA or KIR}

\item{HZY.Red}{Logical Should homozygote genotypes be a single allele for non-DRB345.}
}
\description{
Condenses alleles calls of a single locus string using "+"
}
\note{
This function is for internal use only.
}
