% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAB2GL.R
\name{Tab2GL.Sub}
\alias{Tab2GL.Sub}
\title{Genotype List String Condenser}
\usage{
Tab2GL.Sub(x, System, HZY.Red)
}
\arguments{
\item{x}{Row of loci to condense}

\item{System}{Character Genetic system HLA or KIR}

\item{HZY.Red}{Logical Should homozygote genotypes be a single allele for non-DRB345.}
}
\description{
Condenses column of loci into a GL string using "^"
}
\note{
This function is for internal use only.
}
