% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAB2GL.R
\name{Tab2GL.wrapper}
\alias{Tab2GL.wrapper}
\title{Genotype List String to Tabular Data Conversion}
\usage{
Tab2GL.wrapper(df, System, HZY.Red, Abs.Fill, Cores)
}
\arguments{
\item{df}{Data frame containing GL strings}

\item{System}{Character Genetic system HLA or KIR}

\item{HZY.Red}{Logical Should homozygote genotypes be a single allele for non-DRB345.}

\item{Abs.Fill}{Logical Should absent loci special designations be used}

\item{Cores}{Integer How many cores can be used.}
}
\description{
Expands GL strings to columns of adjacent locus pairs.
}
\note{
This function is for internal use only.
}
