% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{TableMaker}
\alias{TableMaker}
\title{Table Maker}
\usage{
TableMaker(x)
}
\arguments{
\item{x}{Contingency table with binned rare cells.}
}
\description{
Table construction of per haplotype for odds ratio, confidence intervals, and pvalues
}
\note{
This function is for internal BIGDAWG use only.
}
