% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{A}
\alias{A}
\title{Amino Acid Analysis Function}
\usage{
A(loci.ColNames, Locus, genos, grp, nGrp0, nGrp1, ExonAlign, Cores)
}
\arguments{
\item{loci.ColNames}{The column names of the loci being analyzed.}

\item{Locus}{Locus being analyzed.}

\item{genos}{Genotype table.}

\item{grp}{Case/Control or Phenotype groupings.}

\item{nGrp0}{Number of controls.}

\item{nGrp1}{Number of cases.}

\item{ExonAlign}{Exon protein alignment filtered for locus.}

\item{Cores}{Number of cores to use for analysis.}
}
\description{
This is the workhorse function for the amino acid analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
