% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_wrapper.R
\name{A.wrapper}
\alias{A.wrapper}
\title{Amino Acid Wrapper}
\usage{
A.wrapper(
  loci,
  loci.ColNames,
  genos,
  grp,
  Exon,
  EPL,
  Cores,
  Strict.Bin,
  Output,
  Verbose
)
}
\arguments{
\item{loci}{Loci being analyzed.}

\item{loci.ColNames}{The column names of the loci being analyzed.}

\item{genos}{Genotype table.}

\item{grp}{Case/Control or Phenotype groupings.}

\item{Exon}{Exon(s)for targeted analysis.}

\item{EPL}{Protein Alignment List.}

\item{Cores}{Number of cores to use for analysis.}

\item{Strict.Bin}{Logical specify if strict rare cell binning should be used in ChiSq test}

\item{Output}{Data return carryover from main BIGDAWG function}

\item{Verbose}{Summary display carryover from main BIGDAWG function}
}
\description{
Wrapper function for amino acid analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
