% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/A_support_functions.R
\name{AA.df.check}
\alias{AA.df.check}
\title{Contingency Table Check}
\usage{
AA.df.check(x)
}
\arguments{
\item{x}{contingency table.}
}
\description{
Checks amino acid contingency table data frame to ensure required variation exists.
}
\note{
This function is for internal BIGDAWG use only.
}

