% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{cci.pval.list}
\alias{cci.pval.list}
\title{Case Control Odds Ratio Calculation from Epicalc list variation}
\usage{
cci.pval.list(x)
}
\arguments{
\item{x}{List of 2x2 matrices to apply the cci.pvalue function. List output of TableMaker.}
}
\description{
Variation of the cci.pvalue function
}
\note{
This function is for internal BIGDAWG use only.
}
