% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{CheckAlleles}
\alias{CheckAlleles}
\title{HLA Allele Legitimacy Check for Amino Acid Analysis}
\usage{
CheckAlleles(x, y)
}
\arguments{
\item{x}{Exon protein list alignment object.}

\item{y}{Genotypes from data file}
}
\description{
Checks available alleles against data to ensure complete overlap.
}
\note{
This function is for internal BIGDAWG use only.
}
