% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExonPtnList}
\alias{ExonPtnList}
\title{Exon protein alignments.}
\format{
A list where each element is an alignment dataframe for a single locus.
}
\usage{
ExonPtnList
}
\description{
Alignment object for use in the amino acid analysis.
}
\keyword{datasets}
