% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE.R
\name{HWE}
\alias{HWE}
\title{Hardy Weinbergy Equilibrium Function}
\usage{
HWE(Tab, All.ColNames)
}
\arguments{
\item{Tab}{data frame of genotype files post processing.}

\item{All.ColNames}{character vector of Tab object column names.}
}
\description{
This is the main function for the HWE analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
