% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{PreCheck}
\alias{PreCheck}
\title{Data Summary Function}
\usage{
PreCheck(Tab, All.ColNames, rescall, HLA, Verbose, Output)
}
\arguments{
\item{Tab}{Loci available in exon protein list alignment object.}

\item{All.ColNames}{Column names from genotype data.}

\item{rescall}{HLA resolution set for analysis.}

\item{HLA}{HLA BIGDAWG argument passed to function}

\item{Verbose}{Summary display carryover from BIGDAWG function.}

\item{Output}{Data output carryover form BIGDAWG function}
}
\description{
Summary function for sample population within data file.
}
\note{
This function is for internal BIGDAWG use only.
}
