% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{prepData}
\alias{prepData}
\title{Prepare imported data}
\usage{
prepData(Tab)
}
\arguments{
\item{Tab}{Genotypes dataframe.}
}
\description{
Prepare imported data for processing, checks, and analysis.
}
\note{
This function is for internal BIGDAWG use only.
}
