% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general_functions.R
\name{RunChiSq}
\alias{RunChiSq}
\title{Chi-squared Contingency Table Test}
\usage{
RunChiSq(x)
}
\arguments{
\item{x}{Contingency table.}
}
\description{
Calculates chi-squared contingency table tests and bins rare cells.
}
\note{
This function is for internal BIGDAWG use only.
}

